sched: avoid migrating when softint on tgt cpu should be short
The scheduling change (bug 31501544) to avoid putting RT threads on cores that
are handling softint's was catching cases where there was no reason
to believe the softint would take a long time, resulting in unnecessary
migration overhead. This patch reduces the migration to cases where
the core has a softint that is actually likely to take a long time,
as opposed to the RCU, SCHED, and TIMER softints that are rather quick.
Bug: 31752786

Change-Id: Ib4e179f1e15c736b2fdba31070494e357e9fbbe2
(cherry picked from commit 1ea48477565ffe79b40ec0ae23230d13e5141f2f)